+++ internal/common/common.go
@@ -11,6 +11,7 @@ type TestConfig struct {
 	SecurityBundleInstalled      bool
 	GatewayAPISupported          bool
 	IngressNginxSupported        bool
+	ARMNodePoolEnabled           bool
 }
 
 func NewTestConfigWithDefaults() *TestConfig {
@@ -25,12 +26,13 @@ func NewTestConfigWithDefaults() *TestConfig {
 		SecurityBundleInstalled:      true,
 		GatewayAPISupported:          true,
 		IngressNginxSupported:        false,
+		ARMNodePoolEnabled:           false,
 	}
 }
 
 func Run(cfg *TestConfig) {
 	RunApps(cfg)
-	runBasic()
+	runBasic(cfg)
 	runCertManager(cfg.CertManagerSupported)
 	runDNS(cfg.BastionSupported)
 	runMetrics(cfg)